pyproject.toml
[project]
name = "temporal-workflow-explorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "temporalio>=1.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]
